pyproject.toml
[project]
name = "zettleinventorytracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "google-api-python-client>=2.181.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "httpx>=0.28.1",
    "mypy>=1.17.1",
    "pyodbc>=5.3.0",
    "pytz>=2025.2",
    "rich>=14.1.0",
    "ruff>=0.14.5",
    "sqlmodel>=0.0.27",
    "types-pytz>=2025.2.0.20251108",
    "uvicorn>=0.38.0",
]

[tool.mypy]
check_untyped_defs = true
